fix(deepseek-v4): keep MoE routing scores and attention softmax in fp32
Two precision issues that compound across 61 layers and degrade
backbone parity vs reference (observed during MTP parity testing
in #2191):

1. sqrtsoftplus Gate cast routing scores back to bf16 immediately
   after computing sqrt(softplus(x.float())), losing precision for
   expert selection. The HashGate counterpart stays in fp32. Remove
   the .to(scores.dtype) cast so non-hash layers match.

2. eager_attention_with_sink ran softmax in the input dtype (bf16
   under autocast). Force fp32 softmax for numerical stability,
   matching standard practice.

Also fix a stale docstring claiming compress-ratio attention was
not yet implemented — it has been wired in.

+++ nemo_automodel/components/moe/layers.py
@@ -355,7 +355,7 @@ class Gate(nn.Module):
             weights = original_scores.gather(1, indices)
         elif self.score_func == "sqrtsoftplus":
             # sqrt(softplus(x)) = sqrt(log(1 + exp(x))), used in DeepSeek V4.
-            scores = torch.sqrt(F.softplus(scores.float())).to(scores.dtype)
+            scores = torch.sqrt(F.softplus(scores.float()))
             original_scores = scores
 
             if self.e_score_correction_bias is not None:
